fix: refactor code to allow new unit tests
